Enhance OBB head config and predictions
Updated the model configuration for the OBB head by enabling normalization on bounding boxes and centerness predictions. This adjustment aims to improve detection accuracy and robustness by ensuring better alignment of predictions with the object shapes. Additionally, modified the scaling of theta predictions to ensure consistent data types, which promotes smoother integration in downstream processes. These changes should enhance model performance on relevant tasks.

+++ mmdet/models/dense_heads/obb/obb_fcos_head.py
@@ -160,7 +160,7 @@ class OBBFCOSHead(OBBAnchorFreeHead):
             bbox_pred = bbox_pred.exp()
         theta_pred = self.conv_theta(reg_feat)
         if self.scale_theta:
-            theta_pred = self.scale_t(theta_pred)
+            theta_pred = self.scale_t(theta_pred).float()
         return cls_score, bbox_pred, theta_pred, centerness
 
     @force_fp32(apply_to=('cls_scores', 'bbox_preds', 'theta_preds', 'centernesses'))
